Switched label from double type to float

+++ datasets/NyuDepthEncodbleDataset.py
@@ -46,7 +46,7 @@ class NyuDepthEncodableDataset(EncodableDataset):
         img = self.img_preprocessor(Image.open(img_path).convert('RGB'))
         mask = np.array(Image.open(label_path).resize((224, 224)) , dtype=np.float)
         mask /= np.max(mask)
-        mask = torch.tensor(mask)
+        mask = torch.tensor(mask, dtype=torch.float)
 
         # i = img.detach().numpy().transpose(1, 2, 0)
         # plt.figure(0)
